West-Brat Walls, Transit Matrix Fixes, Weird Tile Fix
Adds walls to the North end of the West Brat-Z that were missing from the main Z but not the lower Z, fixes transit matrixes on vault (z6) and the upper-z cabins (z8). Also changes the sprites for the stairs on the underpass to be the regular stairs with alpha 0 instead of the road sprites because it was showing the debug road sprite instead of the actual road when looking down from above.